renamed proxy -> nativeProxy
